+++ assets/javascripts/kitten/components/form/pill-number-input/pill-number-input.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { number } from '@storybook/addon-knobs'
+import { number, boolean } from '@storybook/addon-knobs'
 import { PillNumberInput } from './index'
 
 export const Default = () => (
@@ -7,5 +7,7 @@ export const Default = () => (
     value={number('Value', 1)}
     min={number('Min', 0)}
     max={number('Max', 99)}
+    disableInput={boolean('Disable input?', false)}
+    disabled={boolean('Disable component?', false)}
   />
 )