+++ assets/javascripts/kitten/components/text-copy/stories.js
@@ -13,7 +13,7 @@ const StoryContainer = ({ children }) => (
 )
 
 export default {
-  title: 'TextCopy',
+  title: 'TextCopy/TextCopy',
   component: TextCopy,
 }