+++ assets/javascripts/kitten/components/tours/highlight-halo/stories.js
@@ -14,7 +14,7 @@ const StoryContainer = ({ children }) => (
 )
 
 export default {
-  title: 'Popovers & Tours/HighlightHalo',
+  title: 'Notifications/HighlightHalo',
   component: HighlightHalo,
 }