+++ assets/javascripts/kitten/karl/pages/contribute/payment-choices/payment-choices.stories.js
@@ -4,7 +4,7 @@ import PaymentChoices from './components/payment-choices'
 import { Marger } from '../../../../components/layout/marger'
 
 export default {
-  title: 'Pages/Contribute/PaymentChoices',
+  title: 'pages/Contribute/PaymentChoices',
   component: PaymentChoices,
 }