+++ assets/javascripts/kitten/karl/pages/kissbankers-modal/stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import KissbankersModal from './index'
 
 export default {
-  title: 'Pages/KissbankersPage',
+  title: 'pages/KissbankersPage',
   component: KissbankersModal,
 }